+++ src/SyndicationType/Concrete/TwitterSyndicationType.php
@@ -46,7 +46,7 @@ class TwitterSyndicationType extends AbstractSyndicationType
             sprintf('https://twitter.com/intent/tweet?url=%s&text=%s', rawurlencode($context->getUrl()), rawurlencode($context->getTitle())),
             $this->translator->trans('huh.syndication_type.types.twitter.title'),
             [
-                'class' => 'twitter',
+                'class' => '',
                 'target' => '_blank',
                 'onclick' => 'window.open(this.href,\'\',\'width=500,height=260,modal=yes,left=100,top=50,location=no,menubar=no,resizable=yes,scrollbars=yes,status=no,toolbar=no\');return false',
             ],